+++ docs/.vitepress/theme/Layout.vue
@@ -1,0 +1,14 @@
+<script setup lang="ts">
+import DefaultTheme from 'vitepress/theme';
+import ArticleMeta from './components/ArticleMeta.vue';
+
+const { Layout } = DefaultTheme;
+</script>
+
+<template>
+  <Layout>
+    <template #doc-before>
+      <ArticleMeta />
+    </template>
+  </Layout>
+</template>