+++ docs/.vitepress/theme/index.ts
@@ -1,10 +1,12 @@
 import type { Theme } from 'vitepress';
 import DefaultTheme from 'vitepress/theme';
 import './custom.css';
+import Layout from './Layout.vue';
 import { applyPageMode } from './utils/page-mode';
 
 export default {
   extends: DefaultTheme,
+  Layout,
   enhanceApp({ router }) {
     if (typeof window === 'undefined') {
       return;